test: Fix teams test
